split test command gated dependency tests
