+++ docs/PHASE_PLAN.md
@@ -493,6 +493,9 @@ checked-in docs, tests, and commits only.
 - Legacy `zen build-graph build.zen` validation tests now keep direct and
   transitive gated test dependency rejection coverage in a focused child
   module, aligning the executable graph validation entrypoints.
+- `zen test build.zen` validation tests now keep direct and transitive gated
+  executable dependency rejection coverage in a focused child module, matching
+  the executable graph validation split layout.
 - Resolver-backed behavior impl metadata now builds restored impl-block tasks
   once and reuses them for both impl method signature restoration and omitted
   default-method synthesis. Impl-block declaration collection now also uses a
